move status to api
